use kind_of instead of is_a to workaround a weird typelib behaviour
Typelib reimplements CompoundType#is_a? to return true if the first field of the
structure is of the expected type. This is because, when looking at it from the
C side, the two data structures are equivalent.

This is technical debt, and something that must be changed in Typelib itself.
Until then, use the most common kind_of? accessor to test whether the typelib
value is an enum

It was causing Vizkit to show a structure that starts with an enum field as
a string instead of as a proper tree.